Caso de uso ComprovarPermisos (es un contralador que contiene un método), no haria falta crear un controlador para esto, en caso de que se necesite algun método estático más, creariamos una clase.

+++ app/Http/Controllers/CU_20Controller.php
@@ -5,25 +5,33 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Carpeta;
 use App\Document;
+use Notification;
 
 class CU_20Controller extends Controller {
 
     public function eliminarCarpeta($id) {
+        $permiso=app('App\Http\Controllers\ComprovarPermisos')->comprovarPermis($id);
         $carpeta = Carpeta::find($id);
         $idCarpeta = $carpeta->idCarpetaPare;
-        $carpeta->delete();
-        
-        
-        return redirect('abrirCarpeta/'.$idCarpeta);
+        if (($permiso=='w')||($permiso=='s')){
+            $carpeta->delete();
+            return redirect('abrirCarpeta/'.$idCarpeta);
+        }else{
+            Notification::error("No tens permisos per realitzar aquesta acció.");
+            return redirect('abrirCarpeta/'.$idCarpeta);
+        }
     }
     
     public function eliminarDocumento($id) {
         $document = Document::find($id);
         $idDocument = $document->idCarpeta;
-        $document->delete();
-        
-        
-        return redirect('abrirCarpeta/'.$idDocument);
+        $permiso=app('App\Http\Controllers\ComprovarPermisos')->comprovarPermis($idDocument);//miramos permisos carpeta padre???nose
+        if (($permiso=='w')||($permiso=='s')){
+            $document->delete();
+            return redirect('abrirCarpeta/'.$idDocument);
+        }else{
+            Notification::error("No tens permisos per realitzar aquesta acció.");
+            return redirect('abrirCarpeta/'.$idDocument);
+        }
     }
-
 }
